+++ rollup.config.browser.js
@@ -1,4 +1,8 @@
 import babel from 'rollup-plugin-babel';
+import builtinModules from "builtin-modules"
+
+const pkg = require('./package.json');
+const external = Object.keys(pkg.dependencies).concat(builtinModules);
 
 export default {
     output: {
@@ -11,5 +15,9 @@ export default {
         babel({
             exclude: ['node_modules/**']
         })
-    ]
+    ],
+    external,
+    globals: {
+        'format-date-time': 'format-date-time'
+    }
 };