+++ src/lib/li-log-core.js
@@ -1,3 +1,4 @@
+import DateTimeFormat from 'format-date-time';
 import utils from './utils';
 
 // default browser text styles
@@ -11,6 +12,10 @@ const browserConsoleStyles = {
 
 function Log(userOptions) {
     const { isBrowser, isNode, mergeOptions } = utils;
+    let dateTimeFormatter;
+    if (isNode) {
+        dateTimeFormatter = new DateTimeFormat('HH:mm:ss');
+    }
     const baseOptions = {
         level: 1, // info as default
         coloredOutput: true,
@@ -72,13 +77,20 @@ function Log(userOptions) {
             || (logOptions.outputMethodOnly.length
                 && !logOptions.outputMethodOnly.includes(methodInfo.name))) return;
 
-        let message = `<${methodInfo.name}> ${args}`;
+        let message;
+
+        if (isNode) {
+            message = `${dateTimeFormatter.now()} <${methodInfo.name}> ${args}`;
+        } else {
+            message = `${utils.getTime()} <${methodInfo.name}> ${args}`;
+        }
+
 
         if (logOptions.coloredOutput) {
             if (isBrowser) {
                 message = `%c${message}`;
             } else if (isNode) {
-                message = `|li-log| ${message}`;
+                message = `${message}`;
             }
         }